% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invormal_c_opt.R
\name{invnormal_c_opt}
\alias{invnormal_c_opt}
\title{Critical Value Given a Nominal Error Type I Associated with a Inverse Normal Distribution}
\usage{
invnormal_c_opt(
  alpha = 0.1,
  n = 100,
  theta0 = 1,
  sigma = 1,
  c1 = 0.001,
  c2 = 0.999,
  R = 1000,
  delta = 0.005,
  tolerance = 0.01,
  max_iter = 100
)
}
\arguments{
\item{alpha}{numeric, represents a nominal error type I. Default value is 0.1.}

\item{n}{numeric, represents the size of the sample. Default value is 100.}

\item{theta0}{numeric, represents a location under the null hypothesis of a sample from a Inverse Normal distribution. Default value is 0.5.}

\item{sigma}{numeric, represents the scale parameter of a Inverse Normal distribution. It is assumed known and its default value is 1.}

\item{c1}{numeric, represents a lower bound to the critical value. Default value is 1e-03.}

\item{c2}{numeric, represents an upper bound to the critical value. Default value is 0.99.}

\item{R}{numeric, represents the number of replicates. Default value is 1000.}

\item{delta}{numeric, represents a precision parameter. Default value is 0.005.}

\item{tolerance}{numeric, represents a relative precision with respect a given alpha. Default value is 0.01.}

\item{max_iter}{integer, represents the maximum number of iterations. Default value is 100.}
}
\value{
A list with number of replicates, sample size, nominal error type I, and empirical critical value obtained
associated with a likelihood ratio statistic.
}
\description{
\code{invnormal_c_opt} is used to obtain a critical value to achieve a nominal error type I when we use a random sample from a Inverse Normal distribution.
}
\examples{
# Critical value when we use a random sample of size 40 from a Inverse Normal distribution
# given a nominal error type I equals to 0.1 and R = 80
# to test H_0: theta = 3 vs H_1: theta != 3
invnormal_c_opt(alpha=0.1,n=40,theta0=3,sigma=1,R=80)
}
\references{
Casella, G. and Berger, R. (2003). Statistical Inference, Second Edition. Duxbury Press.

Hogg, R., McKean, J., and Craig, A. (2019) Introduction to Mathematical Statistic.  Eighth edition. Pearson.
}
\author{
Carlos Alberto Cardozo Delgado <cardozorpackages@gmail.com>.
}
